+++ lib/input/socket.go
@@ -1,13 +1,20 @@
 package input
 
 import (
+	"context"
+	"errors"
 	"fmt"
 	"io"
 	"net"
+	"strings"
+	"sync"
+	"time"
 
+	"github.com/Jeffail/benthos/v3/internal/codec"
 	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/lib/input/reader"
 	"github.com/Jeffail/benthos/v3/lib/log"
+	"github.com/Jeffail/benthos/v3/lib/message"
 	"github.com/Jeffail/benthos/v3/lib/metrics"
 	"github.com/Jeffail/benthos/v3/lib/types"
 )
@@ -34,9 +41,10 @@ If the delimiter field is left empty then line feed (\n) is used.`,
 				"unix", "tcp",
 			),
 			docs.FieldCommon("address", "The address to connect to.", "/tmp/benthos.sock", "127.0.0.1:6000"),
-			docs.FieldAdvanced("multipart", "Whether messages should be consumed as multiple parts. If so, each line is consumed as a message parts and the full message ends with an empty line."),
+			codec.ReaderDocs,
+			docs.FieldDeprecated("delimiter"),
+			docs.FieldDeprecated("multipart"),
 			docs.FieldAdvanced("max_buffer", "The maximum message buffer size. Must exceed the largest message to be consumed."),
-			docs.FieldAdvanced("delimiter", "The delimiter to use to detect the end of each message. If left empty line breaks are used."),
 		},
 		Categories: []Category{
 			CategoryNetwork,
@@ -50,8 +58,10 @@ If the delimiter field is left empty then line feed (\n) is used.`,
 type SocketConfig struct {
 	Network   string `json:"network" yaml:"network"`
 	Address   string `json:"address" yaml:"address"`
-	Multipart bool   `json:"multipart" yaml:"multipart"`
+	Codec     string `json:"codec" yaml:"codec"`
 	MaxBuffer int    `json:"max_buffer" yaml:"max_buffer"`
+	// TODO: V4 remove these fields.
+	Multipart bool   `json:"multipart" yaml:"multipart"`
 	Delim     string `json:"delimiter" yaml:"delimiter"`
 }
 
@@ -60,6 +70,7 @@ func NewSocketConfig() SocketConfig {
 	return SocketConfig{
 		Network:   "unix",
 		Address:   "/tmp/benthos.sock",
+		Codec:     "lines",
 		Multipart: false,
 		MaxBuffer: 1000000,
 		Delim:     "",
@@ -70,45 +81,143 @@ func NewSocketConfig() SocketConfig {
 
 // NewSocket creates a new Socket input type.
 func NewSocket(conf Config, mgr types.Manager, log log.Modular, stats metrics.Type) (Type, error) {
-	delim := conf.Socket.Delim
-	if len(delim) == 0 {
-		delim = "\n"
-	}
-	switch conf.Socket.Network {
-	case "tcp", "unix":
-	default:
-		return nil, fmt.Errorf("socket network '%v' is not supported by this input", conf.Socket.Network)
-	}
-	var conn net.Conn
-	rdr, err := reader.NewLines(
-		func() (io.Reader, error) {
-			if conn != nil {
-				conn.Close()
-				conn = nil
-			}
-			var err error
-			conn, err = net.Dial(conf.Socket.Network, conf.Socket.Address)
-			return conn, err
-		},
-		func() {
-			if conn != nil {
-				conn.Close()
-				conn = nil
-			}
-		},
-		reader.OptLinesSetDelimiter(delim),
-		reader.OptLinesSetMaxBuffer(conf.Socket.MaxBuffer),
-		reader.OptLinesSetMultipart(conf.Socket.Multipart),
-	)
+	rdr, err := newSocketClient(conf.Socket, log)
 	if err != nil {
 		return nil, err
 	}
-	return NewAsyncReader(
-		TypeSocket,
-		true,
-		reader.NewAsyncPreserver(rdr),
-		log, stats,
-	)
+	// TODO: Consider removing the async cut off here. It adds an overhead and
+	// we can get the same results by making sure that the async readers forward
+	// CloseAsync all the way through. We would need it to be configurable as it
+	// wouldn't be appropriate for inputs that have real acks.
+	return NewAsyncReader(TypeSocket, true, reader.NewAsyncCutOff(reader.NewAsyncPreserver(rdr)), log, stats)
 }
 
 //------------------------------------------------------------------------------
+
+type socketClient struct {
+	log log.Modular
+
+	conf      SocketConfig
+	codecCtor codec.ReaderConstructor
+
+	codecMut sync.Mutex
+	codec    codec.Reader
+}
+
+func newSocketClient(conf SocketConfig, logger log.Modular) (*socketClient, error) {
+	switch conf.Network {
+	case "tcp", "unix":
+	default:
+		return nil, fmt.Errorf("socket network '%v' is not supported by this input", conf.Network)
+	}
+
+	if len(conf.Delim) > 0 {
+		conf.Codec = "delim:" + conf.Delim
+	}
+	if conf.Multipart && !strings.HasSuffix(conf.Codec, "/multipart") {
+		conf.Codec = conf.Codec + "/multipart"
+	}
+
+	codecConf := codec.NewReaderConfig()
+	codecConf.MaxScanTokenSize = conf.MaxBuffer
+	ctor, err := codec.GetReader(conf.Codec, codecConf)
+	if err != nil {
+		return nil, err
+	}
+
+	return &socketClient{
+		log:       logger,
+		conf:      conf,
+		codecCtor: ctor,
+	}, nil
+}
+
+// ConnectWithContext attempts to establish a connection to the target S3 bucket
+// and any relevant queues used to traverse the objects (SQS, etc).
+func (s *socketClient) ConnectWithContext(ctx context.Context) error {
+	s.codecMut.Lock()
+	defer s.codecMut.Unlock()
+
+	if s.codec != nil {
+		return nil
+	}
+
+	conn, err := net.Dial(s.conf.Network, s.conf.Address)
+	if err != nil {
+		return err
+	}
+
+	if s.codec, err = s.codecCtor("", conn, func(ctx context.Context, err error) error {
+		return nil
+	}); err != nil {
+		conn.Close()
+		return err
+	}
+
+	s.log.Infof("Consuming from socket at '%v://%v'\n", s.conf.Network, s.conf.Address)
+	return nil
+}
+
+// ReadWithContext attempts to read a new message from the target S3 bucket.
+func (s *socketClient) ReadWithContext(ctx context.Context) (types.Message, reader.AsyncAckFn, error) {
+	s.codecMut.Lock()
+	codec := s.codec
+	s.codecMut.Unlock()
+
+	if s.codec == nil {
+		return nil, nil, types.ErrNotConnected
+	}
+
+	parts, codecAckFn, err := codec.Next(ctx)
+	if err != nil {
+		if errors.Is(err, context.Canceled) ||
+			errors.Is(err, context.DeadlineExceeded) {
+			err = types.ErrTimeout
+		}
+		if err != types.ErrTimeout {
+			s.codecMut.Lock()
+			if s.codec != nil && s.codec == codec {
+				s.codec.Close(ctx)
+				s.codec = nil
+			}
+			s.codecMut.Unlock()
+		}
+		if errors.Is(err, io.EOF) {
+			return nil, nil, types.ErrTimeout
+		}
+		return nil, nil, err
+	}
+
+	// We simply bounce rejected messages in a loop downstream so there's no
+	// benefit to aggregating acks.
+	codecAckFn(context.Background(), nil)
+
+	msg := message.New(nil)
+	msg.Append(parts...)
+
+	if msg.Len() == 0 {
+		return nil, nil, types.ErrTimeout
+	}
+
+	return msg, func(rctx context.Context, res types.Response) error {
+		return nil
+	}, nil
+}
+
+// CloseAsync begins cleaning up resources used by this reader asynchronously.
+func (s *socketClient) CloseAsync() {
+	go func() {
+		s.codecMut.Lock()
+		if s.codec != nil {
+			s.codec.Close(context.Background())
+			s.codec = nil
+		}
+		s.codecMut.Unlock()
+	}()
+}
+
+// WaitForClose will block until either the reader is closed or a specified
+// timeout occurs.
+func (s *socketClient) WaitForClose(time.Duration) error {
+	return nil
+}