+++ lib/input/socket_test.go
@@ -2,17 +2,21 @@ package input
 
 import (
 	"errors"
+	"fmt"
 	"net"
 	"reflect"
 	"sync"
 	"testing"
 	"time"
 
+	"github.com/Jeffail/benthos/v3/lib/input/reader"
 	"github.com/Jeffail/benthos/v3/lib/log"
 	"github.com/Jeffail/benthos/v3/lib/message"
 	"github.com/Jeffail/benthos/v3/lib/metrics"
 	"github.com/Jeffail/benthos/v3/lib/response"
 	"github.com/Jeffail/benthos/v3/lib/types"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestSocketBasic(t *testing.T) {
@@ -903,4 +907,144 @@ func TestTCPSocketMultipartShutdown(t *testing.T) {
 	}
 
 	wg.Wait()
+}
+
+func BenchmarkTCPSocketWithCutOff(b *testing.B) {
+	ln, err := net.Listen("tcp", "127.0.0.1:0")
+	if err != nil {
+		ln, err = net.Listen("tcp6", "[::1]:0")
+		require.NoError(b, err)
+	}
+	b.Cleanup(func() {
+		ln.Close()
+	})
+
+	conf := NewConfig()
+	conf.Socket.Network = "tcp"
+	conf.Socket.Address = ln.Addr().String()
+
+	sRdr, err := newSocketClient(conf.Socket, log.Noop())
+	require.NoError(b, err)
+
+	rdr, err := NewAsyncReader(TypeSocket, true, reader.NewAsyncCutOff(reader.NewAsyncPreserver(sRdr)), log.Noop(), metrics.Noop())
+	require.NoError(b, err)
+
+	defer func() {
+		rdr.CloseAsync()
+		assert.NoError(b, rdr.WaitForClose(time.Second))
+	}()
+
+	conn, err := ln.Accept()
+	require.NoError(b, err)
+
+	wg := sync.WaitGroup{}
+	wg.Add(1)
+	go func() {
+		conn.SetWriteDeadline(time.Now().Add(time.Second * 60))
+		for i := 0; i < b.N; i++ {
+			_, cerr := conn.Write([]byte(fmt.Sprintf("hello world this is message %v\n", i)))
+			assert.NoError(b, cerr)
+		}
+		wg.Done()
+	}()
+
+	readNextMsg := func() (string, error) {
+		var payload string
+		select {
+		case tran := <-rdr.TransactionChan():
+			payload = string(tran.Payload.Get(0).Get())
+			go func() {
+				select {
+				case tran.ResponseChan <- response.NewAck():
+				case <-time.After(time.Second):
+				}
+			}()
+		case <-time.After(time.Second):
+			return "", errors.New("timed out")
+		}
+		return payload, nil
+	}
+
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		exp := fmt.Sprintf("hello world this is message %v", i)
+		act, err := readNextMsg()
+		assert.NoError(b, err)
+		assert.Equal(b, exp, act)
+	}
+
+	wg.Wait()
+	conn.Close()
+}
+
+func BenchmarkTCPSocketNoCutOff(b *testing.B) {
+	ln, err := net.Listen("tcp", "127.0.0.1:0")
+	if err != nil {
+		ln, err = net.Listen("tcp6", "[::1]:0")
+		require.NoError(b, err)
+	}
+	b.Cleanup(func() {
+		ln.Close()
+	})
+
+	conf := NewConfig()
+	conf.Socket.Network = "tcp"
+	conf.Socket.Address = ln.Addr().String()
+
+	sRdr, err := newSocketClient(conf.Socket, log.Noop())
+	require.NoError(b, err)
+
+	rdr, err := NewAsyncReader(TypeSocket, true, reader.NewAsyncPreserver(sRdr), log.Noop(), metrics.Noop())
+	require.NoError(b, err)
+
+	defer func() {
+		rdr.CloseAsync()
+		assert.NoError(b, rdr.WaitForClose(time.Second))
+	}()
+
+	conn, err := ln.Accept()
+	require.NoError(b, err)
+
+	wg := sync.WaitGroup{}
+	wg.Add(1)
+	go func() {
+		conn.SetWriteDeadline(time.Now().Add(time.Second * 60))
+		for i := 0; i < b.N; i++ {
+			_, cerr := conn.Write([]byte(fmt.Sprintf("hello world this is message %v\n", i)))
+			assert.NoError(b, cerr)
+		}
+		wg.Done()
+	}()
+
+	readNextMsg := func() (string, error) {
+		var payload string
+		select {
+		case tran := <-rdr.TransactionChan():
+			payload = string(tran.Payload.Get(0).Get())
+			go func() {
+				select {
+				case tran.ResponseChan <- response.NewAck():
+				case <-time.After(time.Second):
+				}
+			}()
+		case <-time.After(time.Second):
+			return "", errors.New("timed out")
+		}
+		return payload, nil
+	}
+
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		exp := fmt.Sprintf("hello world this is message %v", i)
+		act, err := readNextMsg()
+		assert.NoError(b, err)
+		assert.Equal(b, exp, act)
+	}
+
+	wg.Wait()
+	conn.Close()
 }